+++ CHANGES.md
@@ -36,6 +36,7 @@
 - Fixed positioning of batch normalisation in residual mode of `FullyConnected` to after addition
 - `rf_rank_features` was accidentally evaluating feature importance on validation data rather than training data, resulting in lower importances that it should
 - Fixed feature selection in examples using a test size of 0.8 rather than 0.2
+- Fixed crash when no importnat features were found by `rf_rank_features`
 
 ## Changes